refactor(codegen): replace ternary with typeParameterInstantiation attribute in parser
