+++ packages/react-native-codegen/src/parsers/__tests__/parsers-primitives-test.js
@@ -26,6 +26,9 @@ const {
   emitMixedTypeAnnotation,
   typeAliasResolution,
 } = require('../parsers-primitives.js');
+import {MockedParser} from '../parserMock';
+
+const parser = new MockedParser();
 
 describe('emitBoolean', () => {
   describe('when nullable is true', () => {
@@ -334,7 +337,7 @@ describe('typeAliasResolution', () => {
 
 describe('emitPromise', () => {
   const moduleName = 'testModuleName';
-  const language = 'Flow';
+
   describe("when typeAnnotation doesn't have exactly one typeParameter", () => {
     const typeAnnotation = {
       typeParameters: {
@@ -348,7 +351,7 @@ describe('emitPromise', () => {
     it('throws an IncorrectlyParameterizedGenericParserError error', () => {
       const nullable = false;
       expect(() =>
-        emitPromise(moduleName, typeAnnotation, language, nullable),
+        emitPromise(moduleName, typeAnnotation, parser, nullable),
       ).toThrow();
     });
   });
@@ -370,7 +373,7 @@ describe('emitPromise', () => {
         const result = emitPromise(
           moduleName,
           typeAnnotation,
-          language,
+          parser,
           nullable,
         );
         const expected = {
@@ -389,7 +392,7 @@ describe('emitPromise', () => {
         const result = emitPromise(
           moduleName,
           typeAnnotation,
-          language,
+          parser,
           nullable,
         );
         const expected = {
